+++ src/server/repository/layout/widgetRepositoryMock.ts
@@ -1,4 +1,4 @@
-import type { AdjustedWidgetConfig } from "../entities/adjustedWidgetConfig";
+import type { AdjustedWidgetConfig } from "../../entities/adjustedWidgetConfig";
 import type { WidgetRepository } from "./widgetRepository";
 
 export class WidgetRepositoryMock implements WidgetRepository {
